requirements.txt
langchain==0.3.0
langchain-google-genai==2.0.0
langchain-community==0.3.0
pypdf==5.0.0
faiss-cpu==1.8.0
python-dotenv==1.0.1
streamlit==1.39.0
watchdog==5.0.0
sentence-transformers==3.0.0